VAB-600 Springboard BSP version of u-boot
The VAB-600 Springboard has a Board Support Package (BSP) u-boot
released based on the 1.1.4 upstream version. This commit applies
all the changes made to that upstream version to make the BSP version.

All file changes are included and where the BSP has binary files, those
are included for clarity. Checking out this commit should provide
completely the same files as the u-boot source in the BSP.